+++ src/main/java/profect/group1/goormdotcom/payment/controller/external/v1/PaymentApiDocs.java
@@ -63,22 +63,6 @@ public interface PaymentApiDocs {
     );
 
     @Operation(
-            summary = "결제 취소 API",
-            description = "승인된 결제를 취소합니다. 토스에서 발급된 paymentKey(쿼리 파라미터)와 취소 사유/금액 등은 ModelAttribute DTO로 전달됩니다."
-    )
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "성공입니다"),
-            @ApiResponse(
-                    responseCode = "400",
-                    description = "잘못된 요청",
-                    content = @Content(schema = @Schema(implementation = profect.group1.goormdotcom.apiPayload.ApiResponse.class))
-            )
-    })
-    profect.group1.goormdotcom.apiPayload.ApiResponse<PaymentCancelResponseDto> tossPaymentCancel(
-            @ModelAttribute @Valid profect.group1.goormdotcom.payment.controller.external.v1.dto.request.PaymentCancelRequestDto paymentCancelRequestDto
-    );
-
-    @Operation(
             summary = "결제 내역 조회 API (Slice)",
             description = """
             사용자별 결제 내역을 Slice(무한 스크롤) 형태로 조회합니다.<br><br>
