+++ src/test/java/profect/group1/goormdotcom/payment/PaymentServiceTest.java
@@ -1,0 +1,4 @@
+package profect.group1.goormdotcom.payment;
+
+public class PaymentServiceTest {
+}
